Moved out socketinspector
